refactor: articleComment... 파일과 productComment... 파일을 comment...폴더로 생성 및 이동
